Refatoração completa do sistema de chat + correções críticas
