+++ src/Services/ChatService.ts
@@ -9,6 +9,7 @@ const ChatService = {
           usuario: { id: userId },
         },
       });
+
       return response.data.data;
     } catch (error) {
       console.error("Erro ao criar chat:", error);
@@ -19,14 +20,24 @@ const ChatService = {
   async getChats(userId?: number, isAssistant?: boolean) {
     try {
       const base =
-      `/protocolos?fields[0]=ProtocoloID&fields[1]=id&fields[2]=documentId&fields[3]=Status_Finalizado&fields[4]=updatedAt&fields[5]=createdAt&populate[usuario][fields][0]=id&populate[mensagens][fields]=Mensagem,Data_Envio,Leitura&populate[mensagens][populate][remetente][fields]=id,Tipo&sort=updatedAt:desc`;
+        `/protocolos?
+        fields[0]=ProtocoloID
+        &fields[1]=id
+        &fields[2]=documentId
+        &fields[3]=Status_Finalizado
+        &fields[4]=updatedAt
+        &fields[5]=createdAt
+        &populate[usuario][fields][0]=id
+        &populate[mensagens][fields]=Mensagem,Data_Envio,Leitura
+        &populate[mensagens][populate][remetente][fields]=id,Tipo   
+        &sort=updatedAt:desc`;
 
-      const query =
-        isAssistant
-          ? `${base}&filters[Status_Finalizado][$eq]=false`
-          : `${base}&filters[usuario][id][$eq]=${userId}&filters[Status_Finalizado][$eq]=false`;
+      const query = isAssistant
+        ? `${base}&filters[Status_Finalizado][$eq]=false`
+        : `${base}&filters[usuario][id][$eq]=${userId}&filters[Status_Finalizado][$eq]=false`;
 
       const response = await api.get(query);
+
       return response.data.data || [];
     } catch (error) {
       console.error("Erro ao buscar chats:", error);
@@ -34,17 +45,28 @@ const ChatService = {
     }
   },
 
-  async sendMessage(chatId: number, message: string, userId: number, tempId?: string) {
+  async sendMessage(
+    chatId: number,
+    message: string,
+    userId: number
+  ) {
     try {
-      const response = await api.post("/mensagens?populate=remetente", {
-        data: {
-          Mensagem: message,
-          Data_Envio: new Date().toISOString(),
-          protocolo: { id: chatId },
-          remetente: { id: userId },
-          tempId: tempId || null,
-        },
-      });
+      const response = await api.post(
+        "/mensagens?populate=remetente",
+        {
+          data: {
+            Mensagem: message,
+            Data_Envio: new Date().toISOString(),
+            protocolo: {
+              id: chatId,
+            },
+            remetente: {
+              id: userId,
+            },
+          },
+        }
+      );
+
       return response.data.data;
     } catch (error) {
       console.error("Erro ao enviar mensagem:", error);
@@ -57,6 +79,7 @@ const ChatService = {
       const response = await api.get(
         `/mensagens/?filters[protocolo][id][$eq]=${chatId}&populate=*`
       );
+
       return response.data.data || [];
     } catch (error) {
       console.error("Erro ao buscar mensagens:", error);
@@ -66,11 +89,15 @@ const ChatService = {
 
   async endProtocol(chatId: number | string) {
     try {
-      const response = await api.put(`/protocolos/${chatId}`, {
-        data: {
-          Status_Finalizado: true
+      const response = await api.put(
+        `/protocolos/${chatId}`,
+        {
+          data: {
+            Status_Finalizado: true,
+          },
         }
-      });
+      );
+
       return response.data;
     } catch (error) {
       console.error("Erro ao finalizar protocolo:", error);
